# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: fly.ae API
  version: 0.2.0
  description: REST API for guest or email-authenticated PDF upload, processing and secure sharing.
servers:
  - url: http://localhost:8080/api/v1
    description: Local backend
tags:
  - name: Authentication
  - name: Guest sessions
  - name: Categories
  - name: Documents
  - name: Multipart upload
  - name: Sharing
paths:
  /auth/otp/request:
    post:
      tags: [Authentication]
      operationId: requestOtp
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "202":
          description: OTP accepted for delivery regardless of account existence.
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/otp/verify:
    post:
      tags: [Authentication]
      operationId: verifyOtp
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerification"
      responses:
        "200":
          description: OTP consumed and session created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /guest/sessions:
    post:
      tags: [Guest sessions]
      operationId: createGuestSession
      security: []
      description: Creates a short-lived bearer capability for one guest document.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGuestSession"
      responses:
        "201":
          description: Guest session created after legal acceptance.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestSession"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
  /categories:
    get:
      tags: [Categories]
      operationId: listCategories
      security: []
      responses:
        "200":
          description: Active categories in display order.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
  /documents:
    get:
      tags: [Documents]
      operationId: listDocuments
      responses:
        "200":
          description: Current user's non-deleted documents.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Documents]
      operationId: createDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocument"
      responses:
        "201":
          description: Document metadata created in CREATED state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
  /documents/{documentId}:
    parameters:
      - $ref: "#/components/parameters/DocumentId"
    get:
      tags: [Documents]
      operationId: getDocument
      responses:
        "200":
          description: Owner-visible document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Documents]
      operationId: deleteDocument
      responses:
        "204":
          description: Object removed and document marked DELETED.
        "404":
          $ref: "#/components/responses/NotFound"
  /documents/{documentId}/multipart:
    parameters:
      - $ref: "#/components/parameters/DocumentId"
    post:
      tags: [Multipart upload]
      operationId: createMultipartUpload
      responses:
        "201":
          description: Private multipart session created with a one-hour signing window.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultipartSession"
        "409":
          $ref: "#/components/responses/Conflict"
  /documents/{documentId}/multipart/{uploadId}/parts/{partNumber}:
    parameters:
      - $ref: "#/components/parameters/DocumentId"
      - $ref: "#/components/parameters/UploadId"
      - name: partNumber
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          maximum: 10000
    get:
      tags: [Multipart upload]
      operationId: signMultipartPart
      responses:
        "200":
          description: One-hour presigned URL for one upload part.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignedPart"
  /documents/{documentId}/multipart/{uploadId}/complete:
    parameters:
      - $ref: "#/components/parameters/DocumentId"
      - $ref: "#/components/parameters/UploadId"
    post:
      tags: [Multipart upload]
      operationId: completeMultipartUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteMultipart"
      responses:
        "200":
          description: Upload completed, verified and queued for processing.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          $ref: "#/components/responses/BadRequest"
  /documents/{documentId}/multipart/{uploadId}:
    parameters:
      - $ref: "#/components/parameters/DocumentId"
      - $ref: "#/components/parameters/UploadId"
    delete:
      tags: [Multipart upload]
      operationId: abortMultipartUpload
      responses:
        "204":
          description: Multipart upload aborted.
  /shares/{token}:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          minLength: 43
          maxLength: 128
    get:
      tags: [Sharing]
      operationId: resolveShare
      security: []
      responses:
        "200":
          description: Approved document metadata and a short-lived download URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedDocument"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
components:
  securitySchemes:
    bearerSession:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Email session token or a gst_-prefixed guest token.
  parameters:
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UploadId:
      name: uploadId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 512
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Authentication failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: The current principal cannot use this operation.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource was not found or is not visible to the caller.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: Resource state does not allow the operation.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    PayloadTooLarge:
      description: File exceeds the owner-specific upload limit.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    RateLimited:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
  schemas:
    OtpRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
    OtpVerification:
      type: object
      additionalProperties: false
      required:
        [email, code, acceptedLegal, termsVersion, privacyVersion]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        code:
          type: string
          pattern: "^[0-9]{6}$"
        acceptedLegal:
          const: true
        termsVersion:
          type: string
          minLength: 1
          maxLength: 32
        privacyVersion:
          type: string
          minLength: 1
          maxLength: 32
    Session:
      type: object
      required: [accessToken, expiresAt, user]
      properties:
        accessToken:
          type: string
        expiresAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
    CreateGuestSession:
      type: object
      additionalProperties: false
      required: [acceptedLegal, termsVersion, privacyVersion]
      properties:
        acceptedLegal:
          const: true
        termsVersion:
          type: string
          minLength: 1
          maxLength: 32
        privacyVersion:
          type: string
          minLength: 1
          maxLength: 32
    GuestSession:
      type: object
      required: [accessToken, expiresAt, maxFileSizeBytes]
      properties:
        accessToken:
          type: string
          pattern: "^gst_"
        expiresAt:
          type: string
          format: date-time
        maxFileSizeBytes:
          type: integer
          const: 10485760
    User:
      type: object
      required: [id, email]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
    Category:
      type: object
      required: [id, code, name]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
    CreateDocument:
      type: object
      additionalProperties: false
      required: [categoryId, msn, filename, mimeType, sizeBytes]
      properties:
        categoryId:
          type: string
          format: uuid
        msn:
          type: string
          minLength: 1
          maxLength: 64
        filename:
          type: string
          minLength: 5
          maxLength: 255
          pattern: "(?i)^.+\\.pdf$"
        mimeType:
          const: application/pdf
        sizeBytes:
          type: integer
          minimum: 1
          maximum: 104857600
    Document:
      type: object
      required:
        [id, category, msn, filename, mimeType, sizeBytes, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        category:
          $ref: "#/components/schemas/Category"
        msn:
          type: string
        filename:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
          format: int64
        status:
          $ref: "#/components/schemas/DocumentStatus"
        shareUrl:
          type: [string, "null"]
          format: uri
        createdAt:
          type: string
          format: date-time
    DocumentStatus:
      type: string
      enum:
        [CREATED, UPLOADING, PENDING, PROCESSING, APPROVED, REJECTED, FAILED, DELETED]
    MultipartSession:
      type: object
      required: [uploadId, key, expiresAt]
      properties:
        uploadId:
          type: string
        key:
          type: string
        expiresAt:
          type: string
          format: date-time
    SignedPart:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        headers:
          type: object
          additionalProperties:
            type: string
    CompletedPart:
      type: object
      additionalProperties: false
      required: [partNumber, etag]
      properties:
        partNumber:
          type: integer
          minimum: 1
        etag:
          type: string
          minLength: 1
    CompleteMultipart:
      type: object
      additionalProperties: false
      required: [parts]
      properties:
        parts:
          type: array
          minItems: 1
          maxItems: 10000
          items:
            $ref: "#/components/schemas/CompletedPart"
    SharedDocument:
      type: object
      required: [category, msn, filename, sizeBytes, downloadUrl]
      properties:
        category:
          type: string
        msn:
          type: string
        filename:
          type: string
        sizeBytes:
          type: integer
          format: int64
        downloadUrl:
          type: string
          format: uri
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
security:
  - bearerSession: []